Fix typecasting

+++ src/engine/platform/genesis.cpp
@@ -154,7 +154,7 @@ void DivPlatformGenesis::processDAC() {
             if (s->samples>0) {
               while (chan[i].dacPeriod>=(chipClock/576)) {
                 ++chan[i].dacPos;
-                if (!chan[i].dacDirection && (s->isLoopable() && chan[i].dacPos>=(unsigned int)s->EndPosition())) {
+                if (!chan[i].dacDirection && (s->isLoopable() && chan[i].dacPos>=s->EndPosition())) {
                   chan[i].dacPos=s->loopStart;
                 } else if (chan[i].dacPos>=s->samples) {
                   chan[i].dacSample=-1;
@@ -200,7 +200,7 @@ void DivPlatformGenesis::processDAC() {
             }
           }
           chan[5].dacPos++;
-          if (!chan[5].dacDirection && (s->isLoopable() && chan[5].dacPos>=(unsigned int)s->EndPosition())) {
+          if (!chan[5].dacDirection && (s->isLoopable() && chan[5].dacPos>=s->EndPosition())) {
             chan[5].dacPos=s->loopStart;
           } else if (chan[5].dacPos>=s->samples) {
             chan[5].dacSample=-1;
